requirements.txt
flask
flask-mysqldb
flask-wtf
flask-bcrypt
flask-login
python-dotenv
